requirements.txt
Flask==3.0.0
openai>=1.55.0
anthropic>=0.40.0
google-generativeai==0.3.2
numpy<2.0
chromadb==0.4.18
sentence-transformers>=2.6.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
gunicorn==21.2.0
